Authentication Part 2
- remove TorcHTTPServer global authorisation bool
- HTTP
  - allow unauthorised access to methods that do not change state (i.e.
disallow PUT, POST, DELETE if not authorised)
  - request authentication for other methods
- WebSocket
  - allow unauthenticated peers to create WebSockets but disallow any
calls that would require authentication as above.
  - authenticated peers have full access.

N.B. The only reason to retain GetWebSocketToken is to allow web clients
the ability to open an unauthenticated interface and then 'login' by
closing the current websocket and opening a new one after requesting an
access token. The access token will force the web client to request
authentication credentials and thus authenticate the socket.