Major refactor of socket handling
- TorcHTTPConnection needed to be upgraded to a full thread to allow
later handling of secure sockets.
- moving that class into a full thread led to various complexities
around threading.
- far simpler and overally more sensible to move the HTTP handling code
into TorcWebSocket. All sockets are now treated the same and are just
upgraded within the TorcWebSocket code to full websocket status.
- no moving sockets/threads, cleaner access to and handling of the
threads.
- only complication is that TorcNetworkedContext still needs to take
ownership of any threads that it needs to handle as it is just too
convoluted trying to access the thread from another object without it
being deleted while in use.
- removed the abort signal as all sockets now run in their own threads
and thus the event loop can be easily closed.
-